Commit de fin de journée.

+++ Surface.h
@@ -1,10 +1,3 @@
-/*
- * Surface.h
- *
- *  Created on: 14 févr. 2018
- *      Author: postaron
- */
-
 #ifndef SURFACE_H_
 #define SURFACE_H_
 
@@ -14,7 +7,7 @@
 class Surface: public FillProperty, public Geom2D {
 public:
 	virtual ~Surface();
-	virtual float aire() = 0;
+	virtual double aire() = 0;
 	virtual void info() = 0;
 };
